Refactor the PowerCmp function
Why we have this: 
This algorithm here directly follow the rule of ratio to win, and easier to understand. Just think about a smaller space, e.g. 100, instead of 2^256. 
When one miner has 10% power, it should have 10% ratio to win. that is, when the h is in {0..9}, it wins. Simply speaking, it is: 
     (h + 1) / 100 <= 10% ,  instead of     h / 99 < 10%    (both are equivalent when h is an integer and power < 1), 
but the former is easier to understand.