Fix ask Max and Min piece size being reset on reboot
`set-ask` for max and min piece sizes only takes effects while the miner
runs. During reboot they are reset to the defaults.

This is because during initialization the miner does a "default set ask" with
the purpose of adjusting the MaxPieceSize to the miners sector-size.

However this should only happen if the miner has never adjusted the ask
themselves, that is, SeqNo is 0.

Additionally, handle the case when the ask is nil, in case markets code
decides to not create a default ask during initialization at some point.