Testground chatops
This gives you a "slash" command that you can type in issues or pull
requests to submit a task ad-hoc to testground.

You can trigger the testground plan with /testground as the first line
in a comment.

```
/testground plan_directory=testplans/graphsync composition_file=testplans/graphsync/_compositions/stress-k8s.toml

you can include any other information in the comment after the first
line and it will be ignored.
```

If the comment is created on a pull request, the test will be run with
the relevant pull. You can also specify the ref like this:

```
/testground ref=v1.5.3 plan_directory=testplans/graphsync composition_file=testplans/graphsync/_compositions/stress-k8s.toml
```