remove MaxConcurrentReadyFetches=2 default value.
This was there to deal with migration IO workload. This is no
longer a problem with the new migration logic. The initialize-all
operation/command deals with concurrency.

For ongoing operation, miners can set the value they wish, but
we do not make a decision for them by default, because each
deployment can handle different IO loads.