Add used space for all mounts under a storage path
Filesystems like zfs can contain sub-volumes which are distinct datasets
but share free space. There are reasons an operator may want the sealed
or cache directory to be on distinct volumes, which makes the accounting
for the drive capacity inaccurate.

Enumerating all mounts under the storage path, and accumulating
the used space in each accurately determines filesystem capacity.