Eth JSON-RPC API: implement eth_getStorageAt.
This comes with a temporary hack. The EVM runtime actor uses the U256 parameter type in
GetStorageAtParams, which serializes as a hex-encoded string. It should serialize
as bytes. We didn't get to fix in time for Iron, so for now we just pass through the
hex-encoded value passed through the Eth JSON-RPC API, by remarshalling it.

We don't fix this at origin (builtin-actors) because we are not updating the bundle
for Iron.