events: version the MessageReceipt structure.
This is implemented as an internal version discriminator. An alternative
would've been to turn MessageReceipt into an interface with two backing
structs for v0 and v1. But that approach would've broken all usage sites.

CBOR serde for MessageReceipt is no longer automatically generated. We
use custom logic to deal with versioning in both directions. This logic
uses the cborgen serde code as a basis.