Fix dependency caching
The first stab at deps caching didn't work at all. It cached a folder
that was always empty, and the command used to get the dependencies
(make deps) only builds filecoin-ffi and other externs, it doesn't
download the go modules. This now downloads the go modules, and updates
the cache prefix to invalidate the previous caches.