feat: eth: remove support for ETH tracing APIs
Unfortunately, these APIs turned out to be very difficult to correctly
support given all the differences between Filecoin and Ethereum. The
fact of the matter is that Filecoin is not Ethereum. We strive for
Ethereum compatibility from the perspective of _builders_, but that
doesn't mean that the Filecoin chain can be accurately indexed as if it
were an Ethereum chain.

Issues with the Eth-JSON RPC API include:

1. Mapping contract creation semantics from Filecoin to Ethereum
   requires mapping multiple calls into a single "CREATE" operation.
2. Address translation from Filecoin Actor IDs to Ethereum addresses is
   error prone, especially in the face of reverts.
3. Representing native calls in an EVM friendly way is non-trivial.

That's not to say that all this is impossible to fix. However:

1. This code would require significant testing before we'd feel
   comfortable recommending that it be used for indexing.
2. Any indexers built on this code would likely be sub-par.
3. There's no guarantee that we won't make future changes that are
   fundamentally incompatible with Ethereum-style tracing.