fix: events: address sqlite index selection performance regressions
Fixes: https://github.com/filecoin-project/lotus/issues/12255

SQLite was found to be avoiding the intended initial indexes for some types of
complex queries and opting for minor indexes which don't narrow down the search
space enough. Specifically we want queries to first either narrow by height
or tipset_key_cid and then apply other criteria. Having alternative indexes
when a query such as `height>=X AND height<=Y` are encountered cause SQLite to
avoid the height index entirely. By removing additional indexes that could be
used during the main query path (prefillFilter), we force SQLite to use the
intended indexes and narrow the results without the use of indexes.