fix(pubsub): Switch to a last-seen cache strategy
We were using the default first-seen strategy which means we'd loop
messages every 2 minutes (when they expire from the cache). This new
strategy keeps the message in the cache until we haven't seen it for 2
minutes, ensuring that it eventually drops off the network (with high
likelihood).

The correct fix is message expiration:
https://github.com/libp2p/go-libp2p-pubsub/issues/573, but that requires
deeper protocol changes.