Remove build dependency to `goimports` cli
Lotus imports fixing was optimised to run as a go script for a faster
turnaround. The changes here remove the leftover direct calls to
`goimports` such that all imports are now fixed with the same script.
This removes the need for installing `goimports` during CI workflows or
local development environment.

Relates to:
  * https://github.com/filecoin-project/lotus/pull/11695