Cache `make deps` when running CI checks
Over 90% of time running CI check gen is spent on `make deps`. Define a
composite GitHub Action that caches the ffi dependencies and makes this
reusable across other actions.

Note that the changes remove `make lotus` from the gen-check CI workflow
since it is not required for `make gen` to execute.