chore: reimplement `docsgen-cli` in go for faster more maintainable generation (#12510)
* Reimplement docsgen-cli in go for faster more maintainable generation

Replace the python script used to generate markdown documentation from
CLI usage output with a go script. The new implementation runs
approximately 2X faster than the python version.

Ideally, we should use urfav HelpPrinter directly to get the usage in
order to avoid having to build the CLIs first. Based on a quick try it
looks like the usage is not exactly the same as the built version, and
there are env vars ossified deep into the code that changes the usage
printing. Hence the path of least resistance, i.e. re-implementation
in go.